// src/models/response/ObtenerVentaFlashVentasJson.ts
export interface ObtenerVentaFlashVentasJson {
  ID: number;
  NOMBRE: string;
  ciudad: string;
  estado: string;
  director: string;
  supervisor: string;
  ITVTUNIGTN: number;
  ITVTUNIGT: number;
  ITINGTOTN: number;
  ITINGTOT: number;
  ITPRESUP: number;
  ITALCANC: number;
  ALCANCE: number;
  VarVtaPesos: number;
  ITVTUNIMCN: number;
  ITVTUNMCT: number;
  ITVTMCIASN: number;
  ITVTMCIAS: number;
  VARVENTAS: number;
  cuotaMensual: number;
  COBERTURA: number;
  VTAMTLY: number;
  VTAMT: number;
  VARMT: number;
  CostoLY: number;
  CostoCY: number;
  CostoLYORI: number;
  CostoCYORI: number;
  RANKING: number;
  ITVTTAEN: number;
  ITVTTAE: number;
  ITNUTKTOTN: number;
  ITNUTKTOT: number;
  ITNUTKMCIN: number;
  ITNUTKMCI: number;
  ITVTUNTAEN: number;
  ITVTUNTAE: number;
  ITTKPROMCN: number;
  ITTKPROMC: number;
  ITTKPRTAEN: number;
  ITTKPRTAE: number;
  INVOHP: number;
}